refactor: проверки на наличие фильма и/или пользователя вынесена из хранилища в сервис.
